# Conflicts:
#	hni-admin-service/src/main/java/org/hni/admin/filter/JWTTokenAuthenticatingFilter.java
#	hni-admin-service/src/main/java/org/hni/admin/service/UserServiceController.java
#	hni-user/src/main/java/org/hni/user/service/DefaultUserOnboardingService.java

+++ hni-user/src/main/java/org/hni/user/dao/UserPartialCreateDAO.java
@@ -4,5 +4,5 @@ import org.hni.common.dao.BaseDAO;
 import org.hni.user.om.UserPartialData;
 
 public interface UserPartialCreateDAO extends BaseDAO<UserPartialData> {
-	public UserPartialData getUserPartialDataByUserId(int userId);
+	public UserPartialData getUserPartialDataByUserId(Long userId);
 }
