+++ hni-user/src/main/java/org/hni/user/service/DefaultUserPartialCreateService.java
@@ -21,7 +21,7 @@ public class DefaultUserPartialCreateService extends AbstractService<UserPartial
 	}
 
 	@Override
-	public UserPartialData getUserPartialDataByUserId(int userId) {
+	public UserPartialData getUserPartialDataByUserId(Long userId) {
 		return userPartialCreateDAO.getUserPartialDataByUserId(userId);
 	}
